a little bit of prep
changed the size of the generated primes to work with 1024 bit
encryption. Created a build script that shoule work on most linux
distros (stuck with gcc for now to ommit c++ libs I don't plan to use
from being linked in might change later.) increased the buffer size for
the main program input to allow 1024 bit values (and some unneeded
buffer space,) changed the final output to decimal to line up with the
format of the input.